refactor: Convert PrefsStorageBase to Kotlin & use DataStorePrefsStorage
